:bug: fix: avoid deletion if a mapping is referenced by any job and update mapping task url in referenced jobs during update at mapping url field

+++ tofhir-server/src/main/scala/io/tofhir/server/service/MappingService.scala
@@ -2,11 +2,14 @@ package io.tofhir.server.service
 
 import com.typesafe.scalalogging.LazyLogging
 import io.tofhir.engine.model.FhirMapping
+import io.tofhir.server.model.{BadRequest, ResourceNotFound}
+import io.tofhir.server.service.job.IJobRepository
 import io.tofhir.server.service.mapping.IMappingRepository
 
 import scala.concurrent.Future
+import scala.concurrent.ExecutionContext.Implicits.global
 
-class MappingService(mappingRepository: IMappingRepository) extends LazyLogging {
+class MappingService(mappingRepository: IMappingRepository, jobRepository: IJobRepository) extends LazyLogging {
 
   /**
    * Get all mapping metadata from the mapping repository
@@ -38,24 +41,55 @@ class MappingService(mappingRepository: IMappingRepository) extends LazyLogging 
   }
 
   /**
-   * Delete the mapping from the repository
+   * Update the mapping from the repository
+   * If the mapping url is changed, update the mapping url in all jobs that reference this mapping
    * @param projectId project id the mapping belongs to
    * @param id mapping id
    * @param mapping mapping to update
    * @return
    */
   def updateMapping(projectId: String, id: String, mapping: FhirMapping): Future[FhirMapping] = {
-    mappingRepository.putMapping(projectId, id, mapping)
+    // find jobs that reference this mapping and update the mapping url
+    mappingRepository.getMapping(projectId, id).flatMap { oldMapping =>
+      val oldMappingUrl = oldMapping.getOrElse(
+        throw ResourceNotFound("Mapping does not exists.", s"A mapping with id $id does not exists in the mapping repository.")
+      ).url
+      // update each job that references this mapping
+      mappingRepository.putMapping(projectId, id, mapping).flatMap { mapping =>
+        jobRepository.getJobsReferencingMapping(projectId, oldMappingUrl).flatMap { jobs =>
+          Future.sequence(jobs.map { job =>
+            jobRepository.putJob(projectId, job.id, job.copy(
+              // find mapping url and change mapping url in mapping tasks in job
+              mappings = job.mappings.map(mappingTask =>
+                if (mappingTask.mappingRef == oldMappingUrl) mappingTask.copy(mappingRef = mapping.url) else mappingTask)
+            ))
+          }).map(_ => mapping)
+        }
+      }
+
+    }
   }
 
   /**
-   * Delete the mapping from the repository
+   * Delete the mapping from the repository if it is not referenced by any job
    * @param projectId project id the mapping belongs to
    * @param id mapping id
    * @return
    */
   def deleteMapping(projectId: String, id: String): Future[Unit] = {
-    mappingRepository.deleteMapping(projectId, id)
+    // find mapping url from mapping id
+    mappingRepository.getMapping(projectId, id).flatMap { mapping =>
+      //find all jobs that reference this mapping
+      val mappingUrl = mapping.getOrElse(
+        throw ResourceNotFound("Mapping does not exists.", s"A mapping with id $id does not exists in the mapping repository.")
+      ).url
+      jobRepository.getJobsReferencingMapping(projectId, mappingUrl).map { jobs =>
+        if (jobs.isEmpty)
+          mappingRepository.deleteMapping(projectId, id)
+        else
+          throw BadRequest("Mapping is referenced by jobs.", s"The mapping with URL $mappingUrl is referenced by jobs: ${jobs.map(_.id).mkString(", ")}.")
+      }
+    }
   }
 
 }
